Refactor: Organize files into request and response packages
Restructured project to group files into `request` and `response` packages, improving organization and maintainability. Updated imports accordingly to reflect the new structure. No functional changes were made.
