+++ ai-mocks-core/src/commonMain/kotlin/me/kpavlov/aimocks/core/ChatResponseSpecification.kt
@@ -1,6 +1,6 @@
 package me.kpavlov.aimocks.core
 
-import me.kpavlov.mokksy.AbstractResponseDefinition
+import me.kpavlov.mokksy.response.AbstractResponseDefinition
 
 public open class ChatResponseSpecification<T>(
     protected val response: AbstractResponseDefinition<T>,
