+++ mokksy/src/commonMain/kotlin/me/kpavlov/mokksy/Stub.kt
@@ -2,6 +2,13 @@ package me.kpavlov.mokksy
 
 import io.ktor.server.application.ApplicationCall
 import io.ktor.server.response.respond
+import me.kpavlov.mokksy.request.RequestSpecification
+import me.kpavlov.mokksy.response.AbstractResponseDefinition
+import me.kpavlov.mokksy.response.ResponseDefinition
+import me.kpavlov.mokksy.response.SseStreamResponseDefinition
+import me.kpavlov.mokksy.response.StreamResponseDefinition
+import me.kpavlov.mokksy.response.respondWithSseStream
+import me.kpavlov.mokksy.response.respondWithStream
 import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicLong
