+++ mokksy/src/commonTest/kotlin/me/kpavlov/mokksy/RequestSpecificationTest.kt
@@ -20,6 +20,7 @@ import io.mockk.junit5.MockKExtension
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.json.Json
+import me.kpavlov.mokksy.request.RequestSpecification
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.extension.ExtendWith
 import kotlin.test.Test
